+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.2.0] - 2024-06-16
+
+### Added
+- support for linux
+- `list` command to get all aliases
+- clap cli parser
+
 ## [0.1.0] - 2024-05-24
 
 ### Added